backend/requirements.txt
Django
django-extensions
djangorestframework
markdown
django-filter
djangorestframework-simplejwt

